examples/soccer/tactical_requirements.txt
# Additional requirements for tactical analysis system
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
